+++ changelog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unrealeased]
+### Changed
+- Less verbose display of progress
+
 ## [v0.2.0] - 2020-10-31
 ### Added
 - Sample points in the vicinity of the cluster